x Some size_t Warnings
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@886 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/File__Analyze_Buffer.cpp
@@ -104,9 +104,9 @@ void File__Analyze::BS_Begin()
         Element_Offset=Element_Size;
     size_t BS_Size;
     if (Buffer_Offset+Element_Size<=Buffer_Size)
-        BS_Size=(size_t)Element_Size-Element_Offset;
+        BS_Size=(size_t)(Element_Size-Element_Offset);
     else
-        BS_Size=Buffer_Size-(Buffer_Offset+Element_Offset);
+        BS_Size=Buffer_Size-(Buffer_Offset+(size_t)Element_Offset);
     BS->Attach(Buffer+Buffer_Offset+(size_t)Element_Offset, BS_Size);
 }
